final search image upload
Okay. the last one was slightly too big so I went from 100x100px to 75x75px. this should be the sweet spot as the original size I put in was 50x50px and it was too blurry. if this doesn't work I will change the search button to read "search" and we can move on from this. hopefully.