Site title font
I'm trying something out -- I want the font style of the title to match how it looks in the pitch deck I've created so I found the font (for free!) got a license key (jic) and downloaded it. I believe to be able to use it in the file for the header I need to have it saved into the assets folder (under "font" so file structure is *chefs kiss*) and then I can load it in with the @font-face css property. thank you w3 schools